\name{kamus_katadasar}
\alias{kamus_katadasar}
\docType{data}
\title{
Root words dictionary
}
\description{
Dictonary of root words Bahasa Indonesia. It is default dictonary for word stem
retrieval.
}
\usage{data("kamus_katadasar")}

\source{
\url{http://hikaruyuuki.lecture.ub.ac.id/kamus-kata-dasar-dan-stopword-list-bahasa-indonesia/}
}
\examples{
data(kamus_katadasar)
head(kamus_katadasar)
}
\keyword{kata dasar, word stemming}
\seealso{
\code{\link{katadasaR}}, \code{\link{katadasaR-package}}
}
