% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/katadasar.R
\name{katadasar}
\alias{katadasaR}
\alias{katadasar}
\title{Word stemming for Bahasa Indonesia}
\usage{
katadasaR(kata, kamus = NULL)
}
\arguments{
\item{kata}{string vector of length 1, word or token from which word stem is
retrieved.}

\item{kamus}{string vector, additional dictionary to be included.}
}
\description{
Provides a function to retrieve word stem (a.k.a. word stemming)
for Bahasa Indonesia using Nazief and Andriani's algorithm. It consists of
set of functions to remove prefixes, suffixes or both, but still unable for
infixes removal.
This package is ported from C sharp code provided by csharp-indonesia
\url{http://www.csharp-indonesia.com/2014/07/algoritma-stemming-pencarian-kata-dasar.html}.
Credit goes to original author(s).\cr\cr
\code{katadasar} or \code{katadasaR} checks if a word is word stem according
dictionary and do stemming process if it is an affixed word.
}
\examples{
\dontrun{
## Stem one word
  katadasaR("kemanusiaan")

## A set of words
 words <- c("jakarta", "seminar", "penggunaan", "menggurui", "pelajaran", "dimana")
 sapply(words, katadasaR)

}
}
\author{
Nur Andi Setiabudi \email{nurandi.mail@gmail.com}
}
\references{
\url{https://en.wikipedia.org/wiki/Stemming}\cr
\url{http://crpit.scem.westernsydney.edu.au/confpapers/CRPITV38Asian.pdf}
}

